pyproject.toml
[tool.poetry]
name = "learn-rest-framework"
version = "0.1.0"
description = ""
authors = ["HDYS-TTBYS <tatehide151@outlook.jp>"]

[tool.poetry.dependencies]
python = "3.7.9"
Django = "^3.1.5"
djangorestframework = "^3.12.2"
django-cors-headers = "^3.6.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
